CBD-2042: Initial work for build database loader
Still a work in progress, but the current form is able to run through
all the commits in the build-team-manifest and extract all the relevant
information, minus code commits and previous build number, successfully.

A separate manifest package is there as it can be re-used by other
applications, along with the beginnings of a generic Git module in the
utils package.

Change-Id: I604ee34d14af00f4644eaf03111503e36692a862
Reviewed-on: http://review.couchbase.org/83000
Reviewed-by: Kenneth Lareau <kenneth.lareau@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>