CBD-2042: Completed initial version of build DB loader/updater
With this update, the build DB loader is now completed, and with the
addition of keeping track of the last build manifest in the build database,
it's also useable as the updater and able to pick up from interrupted runs
as well.

Tested successfully on the current build-manifests repository, generating
197,969 documents.  Review on selected documents should be done to verify
correctness of the loading/updating.

Change-Id: Ia5528797b96bcc91a8a11232cd086a9f103912d7
Reviewed-on: http://review.couchbase.org/84659
Reviewed-by: Chris Hillery
Tested-by: Kenneth Lareau