CBD-2282: Make the update program into a proper framework
The update program was turned into a proper 'framework' program by allowing
modules to be imported that do the actual changes to the build database.
A 'modules' directory was created to store the various schema changes, along
with a README in the directory on how to create a new module as well as how
to run the update program.

This change requires a new version of python-couchbase-commons.

Change-Id: Iabc9bae1bf34c93b2e76e78686bb3ae20edfab37
Reviewed-on: http://review.couchbase.org/88290
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>