CBD-2284: Add program to regenerate Git repositories for loader
In the case of the checked out bare Git repositories that are used by
the database loader being corrupted or deleted, this program will regen-
erate them from the remote data stored in the project documents.

It simply loops through all the remote information in said documents,
skipping any remote URLs that are no longer accessible.

Change-Id: Ic58c94c285249c7e1d66775a80117d82299c6dc1
Reviewed-on: http://review.couchbase.org/88741
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>