CBD-2550: Add parameter support for build endpoints (for QE)
To handle needed information for QE's 'kickoff-qe' job, parameter support for the build
endpoints (full path only) have been added with a 'filter' parameter that can be set to
a few key values.  An HTTP 400 response is returned if either the parameter set is not
just 'filter' or if the 'filter' value is not one of the valid options.

The concise build endpoints may be supported at a later date.

Change-Id: Id44d8fe718228bcfda49ddcf355c447e554c8f3e
Reviewed-on: http://review.couchbase.org/95708
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>