CBD-2580: Add rate limiting for email notifications
Added a SQLite database backend to keep track of the offline nodes and
stuck jobs, along with a 'checked' count so that email notifications
can be rate limited to prevent 'spam floods'.

Also added a minimum time of 5 minutes for offline nodes and stuck jobs
to remove some false positives.

Change-Id: Ie2932ba6927d0eb65a6f1166f2beb3e4b5f66a49
Reviewed-on: http://review.couchbase.org/96494
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>