CBD-2637: Don't share ccache dirs
It seems that sharing mounted ccache directories between two Docker-based
slaves on the same VM might not work, as the directory for the unit-test
slaves was far larger than the max configured size. Create separate ccache
directories for each slave on a VM now.

Also, update README and inventory and some other cleanup.

Change-Id: I3c2ab6a60d84b217cc6f3a5766ee9049dbeac9b4
Reviewed-on: http://review.couchbase.org/98679
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Kenneth Lareau <kenneth.lareau@couchbase.com>