Ensure Jenkins agents can be recreated by 'Delete' in Jenkins UI
Previously the -retry argument to the Jenkins Swarm jar would cause the
agent to reconnect if it was deleted from Jenkins (up to 5 times). We
would prefer the agent to be discarded and have Docker Swarm create a
fresh one when it is removed from Jenkins. The -noRetryAfterConnected
argument achieves this.

For the moment, only rebuild ubuntu18 and analytics/ubuntu18 - it was
the Analytics team who specifically wanted this feature.

Change-Id: I51a923eafdfad78770b8a57d1a4edcb705504185
Reviewed-on: http://review.couchbase.org/c/build-infra/+/152013
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>