CBD-4127: Add glib2-devel to suse12; remove from amzn2
We added this to the amzn2 image at first to get v8 building on aarch64.
It turns out that it is needed on all Linux platforms to build with
gcc10, regardless of platform or architecture. We don't really want to
add it everywhere only for the v8 build, however, so we hack it in the
v8 build script instead to install this package.

However, this won't work on suse12, because those agents aren't running
on suse12 VMs. So we have to bake it in there.

Change-Id: I20cd0872372ac877b015456b930902e50d7c09f7
Reviewed-on: http://review.couchbase.org/c/build-infra/+/157281
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>