CBD-4424: Configure Linux clamav agent to also serve as database mirror
In order to prevent rate-limiting from running "freshclam" too
frequently, we need to run cvdupdate. It made sense to just run that in
the background on the existing clamav scanning agent since it's already
always around and in the swarm.

Also updated the clamav agent to be based on server-linux-build so it
will be more up-to-date; added a separate "clamav" stack rather than
having clamav as a service in the serverbuild stack; and used Traefik
configs to point our new DNS name clamav-mirror.build.couchbase.com at
mega4 to be served by this agent.

Change-Id: Iaa8b84b529578d85b2c57999e0d1d9de8d49566a
Reviewed-on: http://review.couchbase.org/c/build-infra/+/164034
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>