CBD-4558: Also update amzn2 & unit-test agents for new repo, no swarm jar
Rebuilding the CV agent to use the new repo also pulled in the new
couchbuilder_start.sh script that depends on the newer Jenkins Swarm
plugin, so I had to update that plugin on cv.jenkins. That, in turn,
broke the existing unit-test agents which baked in an older
swarm-client.jar. So now I have to update the Ubuntu20 and Centos7
build agent images, which are also used for the unit-test agents
on cv.jenkins.

Also, we needed to rebuild the aarch64 amzn2 agent to include the new
repo tool to fix Server aarch64 builds. Since this Dockerfile is shared
between aarch64 and x86_64, needed to rebuild that image as well.

Change-Id: I3386e3d02c34d9cd5d8771acfe6a61373b4dd6ce
Reviewed-on: https://review.couchbase.org/c/build-infra/+/168575
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>