CBD-4564: Update OpenSSH on old distros
ssh client on affected images is missing algorithms that are required
for connecting to associated services. This change brings the version
of SSH used by older distros up to current.

Note: Suse 11's OpenSSL is too old to compile newer versions of ssh,
so has also been upgraded to facilitate this change.

Change-Id: I873cf5440f4da4ca274c8b17c2d5510e8796f2a2
Reviewed-on: https://review.couchbase.org/c/build-infra/+/168824
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Blair Watt <blair.watt@couchbase.com>