CBD-4046: Fix up unittest agents on cv.jenkins
They needed tunnel information so they could connect to mega4 instead of
mega2, which previously required setting JENKINS_TUNNEL in the agent
environment. That in turn depended on an update to couchbuilder_start.sh
to honor JENKINS_TUNNEL, which in turn required rebuilding all Jenkins
agents. That's tedious, and moreover, setting JENKINS_TUNNEL everywhere
is tedious and difficult to change again in future.

I found a better workaround: setting the Java system parameter
hudson.TcpSlaveAgentListener.hostName on Jenkins itself, which allows it
to report the correct tunneling information to ALL clients. This is much
cleaner and make changing it in future trivial.

This change adds that parameter and updates both the unittest agents (in
the server-jenkins-buildslaves stack) and the original cv-jenkins-agents
stack to use JENKINS_MASTER=cv.jenkins.couchbase.com (which also should
never need updating in future regardless of where Jenkins moves) and
stop setting JENKINS_TUNNEL.

Change-Id: I4d820d69645063f8ace326ddbf94b62b99ce8cc4
Reviewed-on: https://review.couchbase.org/c/build-infra/+/174171
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>