CBD-5163: Switch SDK centos8 agents to ubi8
Centos8 is no longer supported, and we wanted to have a build and
mock agent based on a supported platform. I did have to use the
centos8 repos for a few packages that RH didn't include in the
UBI repos, however.

For now, only the 'mock' agent is live; I will open a new ticket
to track working with the SDK team to test the ubi8 build agents
and switch to them.

Change-Id: I7d6c0aae2d724bebb3047e96efd2fa5d5f05a065
Reviewed-on: https://review.couchbase.org/c/build-infra/+/181511
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>