Update many Jenkins agents to newer JDK
- After upgrading Jenkins to 2.375.1, a number of agents failed because
they were using a version of Java that didn't support version 55.0 of
the .class file format. Take this opportunity to update a whole bunch
from Java 8 all the way to the latest LTS, OpenJDK 17.0.5+8, and also
eliminate a number of different ways of installing it in favor of a
single unpacking of the Adoptium (Temurin) download. In some cases this
involved creating a new Dockerfile.historic with a wrapper Dockerfile as
the underlying distro was old enough that the existing Dockerfile no
longer built cleanly.

- Introduced JRE_VERSION and friends to couchbase-server/versions,
so at least those agents should be consistent.

- Replaced a couple instances of "slave" with "agent".

- Updated couchbuilder_transition_start.sh to use -passwordFile option.

Change-Id: Id212451fce682981ac81fdfe9a6e5b1a67346bc5
Reviewed-on: https://review.couchbase.org/c/build-infra/+/183665
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>