Ensure 'serverbuild_optcouchbase' volume exists for monster agents
voltron explicit depends on this volume existing, and being mounted
to /opt/couchbase. Due to the change of name of the "serverbuild"
stack recently, it wasn't set up correctly. This should ensure that
it is there regardless of the stack name. (Oddly, 'docker stack
deploy' *will* created a named volume if it doesn't exist, so we
don't need to manually set it up.)

Change-Id: I5d56fbcbf58908f687b9d38123f670824a28b195
Reviewed-on: https://review.couchbase.org/c/build-infra/+/195844
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>