CBD-5604: Fix clang in single linux workers MB-58537: Upgrade git on single-linux workers
When clang (or even a symlink to clang) exists in the same directory as
ld, that is the linker clang will use. This has never been a concern as
we've only ever used a single version of binutils with clang alongside
it in /usr/bin, however now that we are providing a newer binutils in
the worker image, we need to ensure clang isn't landing on the old
version in /usr/bin

This change aims to be as non-breaking as possible. By leaving existing
clang symlinks in place in /usr/bin and creating new clang *and*
binutils symlinks in /usr/local/bin we can ensure:

- existing jobs with hardcoded paths continue to work by landing on the
  tooling in /usr/bin (albeit while using an old linker)
- bare clang calls land in /usr/local/bin with clang9 + new binutils
- absolute path calls to clang-9 or clang-15 work correctly, as ld
  does not exist alongside them so they will instead use the pathed
  binutils tooling in /usr/local/bin

Change-Id: I3ba5c014e9fc5275207ccd0eca97037bc8baee0e
Reviewed-on: https://review.couchbase.org/c/build-infra/+/196865
Tested-by: Blair Watt <blair.watt@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>