CBD-5638: Resolve issues with unit tests
Unit tests were initially configured when we were building on Ubuntu
and Centos, and we had test runners for each platform, now we only
build on Centos, and the platforms we run unit tests on should be
updated to match.

This change removes the existing Ubuntu workers, and scales the
centos7 stack to a single replica, this will still be required for
running 7.0.x unit tests. The remaining workers are replaced with
single-linux workers.

Change-Id: I2060ae5f37bfa42d339ff1e058555ab99e2366e8
Reviewed-on: https://review.couchbase.org/c/build-infra/+/198094
Tested-by: Blair Watt <blair.watt@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>