CBD-5819: Don't overwrite prev_bld_num when updating build database
If the load_build_database job has a failure, it will recover the next
time it is run. In particular, if a build entry for a given build number
already exists in the build database, load_build_database will load that
build entry and then update it, rather than creating a new entry and
overwriting it. However it accidentally would overwrite the prev_bld_num
field in that build entry with the current build-number value for the
corresponding product-version. This could cause the database entry for
that first build to have the same value for "build_num" and
"prev_build_num". This in turn could cause the "commits" field to be
empty since there would be no diffs between the manifests for build_num
and prev_build_num. That, finally, causes jira_commenter to not create
any Jira comments for new commits in that build, since it didn't appear
that there were any new commits.

Solution is simple: if prev_build_num is already set, don't overwrite
it.

Change-Id: Ia74e27147c58d9ec6704fed35f20c6d305b1a34e
Reviewed-on: https://review.couchbase.org/c/build-infra/+/206479
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>