Prefer a "small-and-often" approach to disk clean-up
Prior to CV taking place, we will now run a small handful of
unconditional clean-up tasks:

1. Removing temporary Go directories
2. Cleaning up old docker images

This is followed by conditional clean-up of increasing impact:

1. Test/fuzz caches
2. Module caches
3. Build cache
4. BAZEL clean

Upon failing that, we fall-back to more disruptive measures:

1. Clean-up of all Go caches
2. Removal of Go build cache
3. Full docker prune
4. BAZEL expunge

```sh
A job started hook has been configured by the self-hosted runner administrator
Run '/home/couchbase/disk_cleanup.sh'
  shell: /usr/bin/bash --noprofile --norc -e -o pipefail {0}
Error: Process completed with exit code 1.
```

Now that we're not running this script via `cron`, I've also modified
the logging to output to `stdout` - as well as a logfile - so that we
have some output in GitHub Actions, which is something we ran into
recently.

Change-Id: I040985eb37f77a0f5048e48b136a080115d12fc7
Reviewed-on: https://review.couchbase.org/c/build-infra/+/207698
Tested-by: James Lee <james.lee@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>