jira_commenter: Fix up Jira links in commit messages
Apparently with Cloud Jira, when you post a message that contains a Jira
ticket name in plain text, it replaces it with this complicated little
card in the UI which makes the message very hard to read. Use the same
Wiki markup that we use for GitHub links to avoid this.

Change-Id: Ibfc336575933686a017bcb7c33aaff14ef871f80
Reviewed-on: https://review.couchbase.org/c/build-infra/+/220858
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>