CBD-6145: Back up config files after database
Hub 2024.10.1 only wants to back up to an empty directory, so run the
backup script first and then save the config files.

Also pass `--live-system` to the backup script so it will run without
stopping Hub. This is not necessarily recommended, but it's the way
we've always done it and it doesn't require having scheduled downtime.

Change-Id: If2871611bc344c58150d63c4732a3e40ce5ba470
Reviewed-on: https://review.couchbase.org/c/build-infra/+/222280
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>