Updated cv-swarm inventory and stack; eliminate ubuntu18_docker agents
The ubuntu18_docker agents are running an image that couldn't start on a
number of new swarm nodes due to an old entrypoint script. Ubuntu18 is
long out of support anyway. So I switched the `docker` Jenkins label to
the linux_small agents which had the same hardware setup.

However that caused severe still-unexplained performance penalties for
Operator CV jobs that used that label, so I've removed it again. There
is now a temporary ECS service running the older ubuntu18_docker
configuration, so I've left the `ubuntu18_docker` service with 0
replicas which is referenced by the ECS configuration.

In order to make the linux_small agents come up with `docker.sock`
mounted, I had to modify the `buildx.sh` entrypoint hook because
`/usr/local/bin` apparently wasn't on `$PATH` for the entrypoint, and
that's where `docker` is installed. To minimize risk and get the agents
online as quickly as possible, I created a `:20250224-hack` image tag.
This change includes the `hack` directory showing how the image was
built. It is based from the `:20250107` tag that was running before, and
just modifies the `buildx.sh` hook to explicitly specify
`/usr/local/bin/docker`. This hack should be removed when the agents are
next fully rebuilt.

I deleted the VMs which were running the ubuntu18_docker agents, and
used the space to add a new `linux_large` agent. I also created newer
VMs for `linux_small`, and updated the version of Docker running across
the entire swarm. This change includes the updated inventory.

Change-Id: If5cac0767847b93817b9c6ca342d196a69bfe3c3
Reviewed-on: https://review.couchbase.org/c/build-infra/+/224765
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>