Don't attempt to download Debian ISO
`cdimage.debian.org` only provides the latest version of the ISO, which
means we'd need to update every time they do when we want to create a
template on a new XCP host. These templates are named after the Debian
version, and the corresponding terraform scripting requires a single
template name for the whole cluster.

Instead, just include instructions for placing the ISO in the right
place.

Change-Id: If2a6995127fc04e73ad749f02bcddba24884e585
Reviewed-on: https://review.couchbase.org/c/build-infra/+/226109
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>