Create chroot environments from normal Docker images
Rather than building the packaging chroot environments as part of the
couchbasebuild/server-build image build with `rinse` and `deboostrap`,
we now create them as separate Docker images, and then copy the
resulting filesystem into the builder image. This is easier to maintain,
and isn't susceptible to breakages outside outside our control as when
`rinse` stopped working due to upstream changes in Rocky Linux. The
separate images can also be used by other builder images if needed, or
even used independently.

Change-Id: I2a7f547a9627a92b237877786b220eba64e0091e
Reviewed-on: https://review.couchbase.org/c/build-infra/+/233644
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>