Add DFS-based CV agent image
server-builder-cv adds the following to the base server-builder image:

- GDB (testing crash minidump files)
- Clang (used by numerous kv and magma tests)
- Aspell (used by backup tests)
- XMLSec1 (required by pysaml2, used by ns_server cluster_tests)
- GCovr (used by some kv_engine and magma coverage tests)
- Wget (used by various jobs for downloading things)

Several other tools from the original server-cv image are omitted as
they are either not necessary, or else should be installed via cbdep or
similar dynamic means.

Also tweaks DFS base image to add `chrpath`, and split out the builds of
GMP, MPFR, and MPC so that GDB can be built against them.

Change-Id: I7581a7c23d8e8205c6502b17bfba2121abe1b84a
Reviewed-on: https://review.couchbase.org/c/build-infra/+/234339
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>